Start

+++ src/ru/progwards/alex/Factorial.java
@@ -6,7 +6,7 @@ public class Factorial {
     }
     public static long factorial(long value) {
         long a = 1;
-        for (long i = 1; i <= value; i++) {
+        for (int i = 1; i <= value; i++) {
             a = a * i;
         }
         return a;
